refactor: move DbServerConfigManager to ServerConfig.ts
