+++ src/Core/Discord/VoiceLog/VoiceLog.ts
@@ -4,8 +4,8 @@ import { scheduleJob } from 'node-schedule'
 import Queue from 'promise-queue'
 import type { ILogObj, Logger } from 'tslog'
 import { instances } from '../../../Utils/Instances'
-import { type DbServerConfigManager, ERR_DB_NOT_INIT } from '../../SQLite/Core'
-import { type IServerConfig, VoiceMessageTTSType } from '../../SQLite/db/ServerConfig'
+import { ERR_DB_NOT_INIT } from '../../SQLite/Core'
+import { type DbServerConfigManager, type IServerConfig, VoiceMessageTTSType } from '../../SQLite/db/ServerConfig'
 import type { Discord } from '../Core'
 import { VoiceLogCommands } from './VoiceLog/Commands'
 import { VoiceLogText } from './VoiceLog/Text'
