+++ src/Core/Discord/VoiceLog/VoiceLog/Commands.ts
@@ -4,8 +4,7 @@ import type { CommandContext, MessageEmbedOptions } from 'slash-create'
 import { vsprintf } from 'sprintf-js'
 import type { ILogObj, Logger } from 'tslog'
 import { instances } from '../../../../Utils/Instances'
-import type { DbServerConfigManager } from '../../../SQLite/Core'
-import type { IVoiceMessageTTS, VoiceMessageTTSType } from '../../../SQLite/db/ServerConfig'
+import type { DbServerConfigManager, IVoiceMessageTTS, VoiceMessageTTSType } from '../../../SQLite/db/ServerConfig'
 import type { Discord } from '../../Core'
 import type { VoiceLog } from '../VoiceLog'
 import { VoiceLogSetStatus } from './Text'
