+++ src/Core/Discord/VoiceLog/VoiceLog/Voice.ts
@@ -6,7 +6,7 @@ import Queue from 'promise-queue'
 import type { CommandContext, MessageEmbedOptions } from 'slash-create'
 import type { ILogObj, Logger } from 'tslog'
 import { instances } from '../../../../Utils/Instances'
-import type { DbServerConfigManager } from '../../../SQLite/Core'
+import type { DbServerConfigManager } from '../../../SQLite/db/ServerConfig'
 import type { Discord } from '../../Core'
 import { DiscordVoice } from '../../Core/Voice'
 import type { VoiceLog } from '../VoiceLog'
